append bool and heredoc signal fix

+++ srcs/execute/redirection.c
@@ -25,41 +25,50 @@ int	get_input(char *input_file)
 
 void	signal_from_heredoc(int signal)
 {
-	char c = 4;
 
-	ft_dprintf(STDOUT_FILENO, "heredoc hanlder\n");
+	//ft_dprintf(STDOUT_FILENO, "heredoc handler\n");
 	if (signal == SIGINT)
-		write (STDOUT_FILENO, &c, 1);
-	if (signal == SIGQUIT)
-		ft_dprintf(STDOUT_FILENO, "Quit: %d\n", signal);
+		exit (signal);
+	// if (signal == SIGQUIT)
+	// 	ft_dprintf(STDOUT_FILENO, "Quit: %d\n", signal);
 }
 
 int	read_input_heredoc(char *delimiter)
 {
 	char	*line;
 	int		input_fd;
+	pid_t	pid;
 
 	line = NULL;
 	input_fd = open("/tmp/heredoc", O_CREAT | O_RDWR | O_TRUNC , 0644);
 	if (input_fd == -1)
 	system_error("input file does not exist");
-	signal(SIGINT, signal_from_heredoc);
-	signal(SIGQUIT, signal_from_heredoc);
-	while (true)
+	//signal(SIGQUIT, signal_from_heredoc);
+	pid = fork();
+	if (pid == -1)
+		system_error("Fork Error");
+	if (pid == 0)
 	{
-		line = readline("> ");
-		if (line == NULL || ft_streql(line, delimiter))
-			break;
-		ft_putstr_fd (line, input_fd);
-		ft_putstr_fd ("\n", input_fd);
-		free (line);
+		signal(SIGINT, signal_from_heredoc);
+		while (true)
+		{
+			line = readline("> ");
+			if (line == NULL || ft_streql(line, delimiter))
+			{
+				free (line);
+				close (input_fd);
+				input_fd = get_input("/tmp/heredoc");
+				unlink ("/tmp/heredoc");
+				break;
+			}
+			ft_putstr_fd (line, input_fd);
+			ft_putstr_fd ("\n", input_fd);
+			free (line);
+		}
 	}
-	free (line);
-	close (input_fd);
-	input_fd = get_input("/tmp/heredoc");
-	unlink ("/tmp/heredoc");
-	signal(SIGINT, received_signal);
-	signal(SIGQUIT, received_signal);
+	signal(SIGINT, signal_from_child);
+	wait (0);
+	//signal(SIGQUIT, received_signal);
 	return (input_fd);
 }
